requirements.txt
biopython==1.85
bokeh==3.7.3
certifi==2025.7.14
charset-normalizer==3.4.2
checkm-genome==1.2.4
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
dask==2025.5.1
dask-glm==0.3.2
dask-ml==2025.1.0
DendroPy==5.0.8
distributed==2025.5.1
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
gunc==1.0.6
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
locket==1.0.0
lz4==4.4.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.0
multipledispatch==1.0.0
narwhals==1.41.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
partd==1.4.2
pillow==11.2.1
plotly==6.2.0
psutil==7.0.0
pyarrow==20.0.0
pyparsing==3.2.3
pysam==0.23.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
sparse==0.17.0
sympy==1.14.0
tblib==3.1.0
threadpoolctl==3.6.0
toolz==1.0.0
torch==2.7.1
tornado==6.5.1
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
xyzservices==2025.4.0
zict==3.0.0
